attempting to set up pathfinding.
I have the ground and wall tiles set up - setting up the logic for Hero to obtin target position. Thinking go to Chest location if he does not see player. then in a few second stops goes idle and enters a looking around mode to look for the player and repeat.